requirements.txt
Django==3.1.5
django-debug-toolbar==1.10.1
python-decouple==3.1
pytz==2018.5
sqlparse==0.2.4
Pillow==8.1.0
django-allauth
django-crispy-forms
django-bootstrap4
django-bootstrap4
bootstrap4
django-ckeditor
